Work in progress (Mission and Absolut location)
