--- .idea/modules.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/9TekAS-work_9_all_teni-rosettadrone.iml" filepath="$PROJECT_DIR$/9TekAS-work_9_all_teni-rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/app/9TekAS-work_9_all_teni-rosettadrone-app.iml" filepath="$PROJECT_DIR$/app/9TekAS-work_9_all_teni-rosettadrone-app.iml" group="rosettadrone/app" />
-      <module fileurl="file://$PROJECT_DIR$/Documents-9TekAS-work_9_all_teni-rosettadrone.iml" filepath="$PROJECT_DIR$/Documents-9TekAS-work_9_all_teni-rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/Documents-9TekAS-work_9_all_teni-rosettadrone~1.iml" filepath="$PROJECT_DIR$/Documents-9TekAS-work_9_all_teni-rosettadrone~1.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/Linux-rosettadrone.iml" filepath="$PROJECT_DIR$/Linux-rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/app/Linux-rosettadrone-app.iml" filepath="$PROJECT_DIR$/app/Linux-rosettadrone-app.iml" group="rosettadrone/app" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" group="rosettadrone/app" />
-      <module fileurl="file://$PROJECT_DIR$/dji-Linux-rosettadrone.iml" filepath="$PROJECT_DIR$/dji-Linux-rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/rosettadrone.iml" filepath="$PROJECT_DIR$/rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/app/rosettadrone-app.iml" filepath="$PROJECT_DIR$/app/rosettadrone-app.iml" group="rosettadrone/app" />
-      <module fileurl="file://$PROJECT_DIR$/work_9_all_teni-rosettadrone.iml" filepath="$PROJECT_DIR$/work_9_all_teni-rosettadrone.iml" group="rosettadrone" />
-      <module fileurl="file://$PROJECT_DIR$/app/work_9_all_teni-rosettadrone-app.iml" filepath="$PROJECT_DIR$/app/work_9_all_teni-rosettadrone-app.iml" group="rosettadrone/app" />
-    </modules>
-  </component>
-</project>
