+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 28
     defaultConfig {
         applicationId "sq.rogue.rosettadrone"
-        minSdkVersion 25
+        minSdkVersion 22
         targetSdkVersion 28
         versionCode 1
         versionName "5.0.1"
