+++ app/src/main/AndroidManifest.xml
@@ -85,7 +85,7 @@
             android:label="Settings">
         </activity>
 
-        <activity android:name="sq.rogue.rosettadrone.settings.HelpActivity"
+        <activity android:name="sq.rogue.rosettadrone.settings.MapActivity"
             android:theme="@style/AppTheme.Settings"
             android:label="Settings">
         </activity>
