+++ app/src/main/java/sq/rogue/rosettadrone/settings/AiActivity.java
@@ -1,0 +1,61 @@
+package sq.rogue.rosettadrone.settings;
+
+import android.os.Bundle;
+import android.util.Log;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.PreferenceScreen;
+import sq.rogue.rosettadrone.R;
+import sq.rogue.rosettadrone.settings.drone.DroneSettingsFragment;
+
+public class AiActivity extends AppCompatActivity implements PreferenceFragmentCompat.OnPreferenceStartScreenCallback {
+    private static final String TAG = AiActivity.class.getSimpleName();
+
+    /**
+     * @param savedInstanceState Any saved state we are carrying over into the new activity instance
+     */
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        Log.d(TAG, "onCreate");
+
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_help);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setDisplayShowHomeEnabled(true);
+
+//        // Display the fragment as the main content.
+//        getFragmentManager().beginTransaction()
+//                .replace(android.R.id.content, new SettingsFragment())
+//                .commit();
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        onBackPressed();
+        return true;
+    }
+
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
+        setTitle(R.string.settings);
+    }
+
+    @Override
+    public boolean onPreferenceStartScreen(PreferenceFragmentCompat caller, PreferenceScreen preferenceScreen) {
+        Log.d(TAG, "onPreferenceStartScreen");
+//        Log.d(TAG, "callback called to attach the preference sub screen");
+        FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
+        DroneSettingsFragment fragment = new DroneSettingsFragment();
+        Bundle args = new Bundle();
+        args.putString(PreferenceFragmentCompat.ARG_PREFERENCE_ROOT, preferenceScreen.getKey());
+        fragment.setArguments(args);
+        ft.replace(R.id.pref_container, fragment, preferenceScreen.getKey());
+        ft.addToBackStack(preferenceScreen.getKey());
+        ft.commit();
+        setTitle(preferenceScreen.getTitle());
+        return true;
+    }
+}
