--- app/src/main/java/sq/rogue/rosettadrone/settings/HelpActivity.java
@@ -1,677 +1,0 @@
-package sq.rogue.rosettadrone.settings;
-
-import android.Manifest;
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.os.Build;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.LinearLayout;
-import android.widget.RadioGroup;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.google.android.gms.maps.CameraUpdate;
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.BitmapDescriptorFactory;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.Marker;
-import com.google.android.gms.maps.model.MarkerOptions;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.fragment.app.FragmentActivity;
-import androidx.fragment.app.FragmentTransaction;
-import androidx.preference.PreferenceFragmentCompat;
-import androidx.preference.PreferenceScreen;
-import dji.common.error.DJIError;
-import dji.common.flightcontroller.FlightControllerState;
-import dji.common.mission.waypoint.Waypoint;
-import dji.common.mission.waypoint.WaypointMission;
-import dji.common.mission.waypoint.WaypointMissionDownloadEvent;
-import dji.common.mission.waypoint.WaypointMissionExecutionEvent;
-import dji.common.mission.waypoint.WaypointMissionFinishedAction;
-import dji.common.mission.waypoint.WaypointMissionFlightPathMode;
-import dji.common.mission.waypoint.WaypointMissionHeadingMode;
-import dji.common.mission.waypoint.WaypointMissionUploadEvent;
-import dji.common.useraccount.UserAccountState;
-import dji.common.util.CommonCallbacks;
-import dji.sdk.base.BaseProduct;
-import dji.sdk.flightcontroller.FlightController;
-import dji.sdk.mission.waypoint.WaypointMissionOperator;
-import dji.sdk.mission.waypoint.WaypointMissionOperatorListener;
-import dji.sdk.products.Aircraft;
-import dji.sdk.sdkmanager.DJISDKManager;
-import dji.sdk.useraccount.UserAccountManager;
-import sq.rogue.rosettadrone.DJISimulatorApplication;
-import sq.rogue.rosettadrone.R;
-import sq.rogue.rosettadrone.RDApplication;
-import sq.rogue.rosettadrone.settings.drone.DroneSettingsFragment;
-
-//public class HelpActivity extends AppCompatActivity implements OnMapReadyCallback{
-//public class HelpActivity extends AppCompatActivity implements PreferenceFragmentCompat.OnPreferenceStartScreenCallback, OnMapReadyCallback{
-public class HelpActivity extends FragmentActivity implements View.OnClickListener, GoogleMap.OnMapClickListener, OnMapReadyCallback {
-    private static final String TAG = HelpActivity.class.getSimpleName();
-
-    private GoogleMap gMap;
-
-    private Button locate, add, clear, exit;
-    private Button config, upload, start, stop;
-
-    private boolean isAdd = false;
-
-    private double droneLocationLat = 30, droneLocationLng = 60;
-    private final Map<Integer, Marker> mMarkers = new ConcurrentHashMap<Integer, Marker>();
-    private Marker droneMarker = null;
-
-    private float altitude = 100.0f;
-    private float mSpeed = 10.0f;
-
-    private List<Waypoint> waypointList = new ArrayList<>();
-
-    public static WaypointMission.Builder waypointMissionBuilder;
-    private FlightController mFlightController;
-    private WaypointMissionOperator instance;
-    private WaypointMissionFinishedAction mFinishedAction = WaypointMissionFinishedAction.NO_ACTION;
-    private WaypointMissionHeadingMode mHeadingMode = WaypointMissionHeadingMode.AUTO;
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        initFlightController();
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-    }
-
-    @Override
-    protected void onDestroy() {
-        if(mReceiver.getResultCode() < 0 )
-            unregisterReceiver(mReceiver);
-        removeListener();
-        super.onDestroy();
-    }
-
-    /**
-     * @Description : RETURN Button RESPONSE FUNCTION
-     */
-    public void onReturn(View view) {
-        Log.d(TAG, "onReturn");
-        this.finish();
-    }
-
-    private void setResultToToast(final String string) {
-        HelpActivity.this.runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                Toast.makeText(HelpActivity.this, string, Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
-    private void initUI() {
-
-        locate = (Button) findViewById(R.id.locate);
-        add = (Button) findViewById(R.id.add);
-        clear = (Button) findViewById(R.id.clear);
-        config = (Button) findViewById(R.id.config);
-        exit = (Button) findViewById(R.id.exit);
-        upload = (Button) findViewById(R.id.upload);
-        start = (Button) findViewById(R.id.start);
-        stop = (Button) findViewById(R.id.stop);
-
-        locate.setOnClickListener(this);
-        add.setOnClickListener(this);
-        clear.setOnClickListener(this);
-        config.setOnClickListener(this);
-        exit.setOnClickListener(this);
-        upload.setOnClickListener(this);
-        start.setOnClickListener(this);
-        stop.setOnClickListener(this);
-
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        // When the compile and target version is higher than 22, please request the
-        // following permissions at runtime to ensure the
-        // SDK work well.
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            ActivityCompat.requestPermissions(this,
-                    new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.VIBRATE,
-                            Manifest.permission.INTERNET, Manifest.permission.ACCESS_WIFI_STATE,
-                            Manifest.permission.WAKE_LOCK, Manifest.permission.ACCESS_COARSE_LOCATION,
-                            Manifest.permission.ACCESS_NETWORK_STATE, Manifest.permission.ACCESS_FINE_LOCATION,
-                            Manifest.permission.CHANGE_WIFI_STATE, Manifest.permission.MOUNT_UNMOUNT_FILESYSTEMS,
-                            Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.SYSTEM_ALERT_WINDOW,
-                            Manifest.permission.READ_PHONE_STATE,
-                    }
-                    , 1);
-        }
-
-        setContentView(R.layout.activity_help);
-
-    //    IntentFilter filter = new IntentFilter();
-    //    filter.addAction(DJIDemoApplication.FLAG_CONNECTION_CHANGE);
-    //    registerReceiver(mReceiver, filter);
-
-        initUI();
-
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
-                .findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-
-        addListener();
-    }
-
-    protected BroadcastReceiver mReceiver = new BroadcastReceiver() {
-
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            onProductConnectionChange();
-        }
-    };
-
-    private void onProductConnectionChange() {
-        initFlightController();
-        loginAccount();
-    }
-
-    private void loginAccount() {
-
-        UserAccountManager.getInstance().logIntoDJIUserAccount(this,
-                new CommonCallbacks.CompletionCallbackWith<UserAccountState>() {
-                    @Override
-                    public void onSuccess(final UserAccountState userAccountState) {
-                        Log.e(TAG, "Login Success");
-                    }
-
-                    @Override
-                    public void onFailure(DJIError error) {
-                        setResultToToast("Login Error:"
-                                + error.getDescription());
-                    }
-                });
-    }
-
-    private void initFlightController() {
-
-        BaseProduct product;
-
-        if (RDApplication.getSim() == true) {
-            product = DJISimulatorApplication.getAircraftInstance();
-        }else {
-            product = RDApplication.getProductInstance();
-        }
-
-        if (product != null && product.isConnected()) {
-            if (product instanceof Aircraft) {
-                mFlightController = ((Aircraft) product).getFlightController();
-            }
-        }
-
-        if (mFlightController != null) {
-            mFlightController.setStateCallback(djiFlightControllerCurrentState -> {
-                droneLocationLat = djiFlightControllerCurrentState.getAircraftLocation().getLatitude();
-                droneLocationLng = djiFlightControllerCurrentState.getAircraftLocation().getLongitude();
-                updateDroneLocation();
-            });
-        }
-    }
-
-    //Add Listener for WaypointMissionOperator
-    private void addListener() {
-        if (getWaypointMissionOperator() != null) {
-            getWaypointMissionOperator().addListener(eventNotificationListener);
-        }
-    }
-
-    private void removeListener() {
-        if (getWaypointMissionOperator() != null) {
-            getWaypointMissionOperator().removeListener(eventNotificationListener);
-        }
-    }
-
-    private WaypointMissionOperatorListener eventNotificationListener = new WaypointMissionOperatorListener() {
-        @Override
-        public void onDownloadUpdate(WaypointMissionDownloadEvent downloadEvent) {
-
-        }
-
-        @Override
-        public void onUploadUpdate(WaypointMissionUploadEvent uploadEvent) {
-
-        }
-
-        @Override
-        public void onExecutionUpdate(WaypointMissionExecutionEvent executionEvent) {
-
-        }
-
-        @Override
-        public void onExecutionStart() {
-
-        }
-
-        @Override
-        public void onExecutionFinish(@Nullable final DJIError error) {
-            setResultToToast("Execution finished: " + (error == null ? "Success!" : error.getDescription()));
-        }
-    };
-
-    public WaypointMissionOperator getWaypointMissionOperator() {
-        if (instance == null) {
-            if (DJISDKManager.getInstance().getMissionControl() != null) {
-                instance = DJISDKManager.getInstance().getMissionControl().getWaypointMissionOperator();
-            }
-        }
-        return instance;
-    }
-
-    private void setUpMap() {
-        gMap.setOnMapClickListener(this);// add the listener for click for amap object
-
-    }
-
-    @Override
-    public void onMapClick(LatLng point) {
-        if (isAdd == true) {
-            markWaypoint(point);
-            Waypoint mWaypoint = new Waypoint(point.latitude, point.longitude, altitude);
-            //Add Waypoints to Waypoint arraylist;
-            if (waypointMissionBuilder != null) {
-                waypointList.add(mWaypoint);
-                waypointMissionBuilder.waypointList(waypointList).waypointCount(waypointList.size());
-            } else {
-                waypointMissionBuilder = new WaypointMission.Builder();
-                waypointList.add(mWaypoint);
-                waypointMissionBuilder.waypointList(waypointList).waypointCount(waypointList.size());
-            }
-        } else {
-            setResultToToast("Cannot Add Waypoint");
-        }
-    }
-
-    public static boolean checkGpsCoordination(double latitude, double longitude) {
-        return (latitude > -90 && latitude < 90 && longitude > -180 && longitude < 180) && (latitude != 0f && longitude != 0f);
-    }
-
-    // Update the drone location based on states from MCU.
-    private void updateDroneLocation() {
-
-        LatLng pos = new LatLng(droneLocationLat, droneLocationLng);
-        //Create MarkerOptions object
-        final MarkerOptions markerOptions = new MarkerOptions();
-        markerOptions.position(pos);
-        markerOptions.icon(BitmapDescriptorFactory.fromResource(R.drawable.aircraft));
-
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                if (droneMarker != null) {
-                    droneMarker.remove();
-                }
-
-                if (checkGpsCoordination(droneLocationLat, droneLocationLng)) {
-                    droneMarker = gMap.addMarker(markerOptions);
-                }
-            }
-        });
-    }
-
-    private void markWaypoint(LatLng point) {
-        //Create MarkerOptions object
-        MarkerOptions markerOptions = new MarkerOptions();
-        markerOptions.position(point);
-        markerOptions.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_BLUE));
-        Marker marker = gMap.addMarker(markerOptions);
-        mMarkers.put(mMarkers.size(), marker);
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.locate: {
-                updateDroneLocation();
-                cameraUpdate(); // Locate the drone's place
-                break;
-            }
-            case R.id.add: {
-                enableDisableAdd();
-                break;
-            }
-            case R.id.clear: {
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        gMap.clear();
-                    }
-
-                });
-                waypointList.clear();
-                if (waypointMissionBuilder != null) {
-                    waypointMissionBuilder.waypointList(waypointList);
-                }
-                updateDroneLocation();
-                break;
-            }
-            case R.id.config: {
-                showSettingDialog();
-                break;
-            }
-            case R.id.upload: {
-                uploadWayPointMission();
-                break;
-            }
-            case R.id.start: {
-                startWaypointMission();
-                break;
-            }
-            case R.id.stop: {
-                stopWaypointMission();
-                break;
-            }
-            case R.id.exit: {
-                super.onBackPressed();
-                break;
-            }
-            default:
-                break;
-        }
-    }
-
-    private void cameraUpdate() {
-        LatLng pos = new LatLng(droneLocationLat, droneLocationLng);
-        float zoomlevel = (float) 18.0;
-        CameraUpdate cu = CameraUpdateFactory.newLatLngZoom(pos, zoomlevel);
-        gMap.moveCamera(cu);
-    }
-  
-    private void enableDisableAdd() {
-        if (isAdd == false) {
-            isAdd = true;
-            add.setText("Exit");
-        } else {
-            isAdd = false;
-            add.setText("Add");
-        }
-    }
-
-    private void showSettingDialog() {
-        LinearLayout wayPointSettings = (LinearLayout) getLayoutInflater().inflate(R.layout.dialog_waypointsetting, null);
-
-        final TextView wpAltitude_TV = (TextView) wayPointSettings.findViewById(R.id.altitude);
-        RadioGroup speed_RG = (RadioGroup) wayPointSettings.findViewById(R.id.speed);
-        RadioGroup actionAfterFinished_RG = (RadioGroup) wayPointSettings.findViewById(R.id.actionAfterFinished);
-        RadioGroup heading_RG = (RadioGroup) wayPointSettings.findViewById(R.id.heading);
-
-        speed_RG.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
-
-            @Override
-            public void onCheckedChanged(RadioGroup group, int checkedId) {
-                if (checkedId == R.id.lowSpeed) {
-                    mSpeed = 3.0f;
-                } else if (checkedId == R.id.MidSpeed) {
-                    mSpeed = 5.0f;
-                } else if (checkedId == R.id.HighSpeed) {
-                    mSpeed = 10.0f;
-                }
-            }
-
-        });
-
-        actionAfterFinished_RG.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
-
-            @Override
-            public void onCheckedChanged(RadioGroup group, int checkedId) {
-                Log.d(TAG, "Select finish action");
-                if (checkedId == R.id.finishNone) {
-                    mFinishedAction = WaypointMissionFinishedAction.NO_ACTION;
-                } else if (checkedId == R.id.finishGoHome) {
-                    mFinishedAction = WaypointMissionFinishedAction.GO_HOME;
-                } else if (checkedId == R.id.finishAutoLanding) {
-                    mFinishedAction = WaypointMissionFinishedAction.AUTO_LAND;
-                } else if (checkedId == R.id.finishToFirst) {
-                    mFinishedAction = WaypointMissionFinishedAction.GO_FIRST_WAYPOINT;
-                }
-            }
-        });
-
-        heading_RG.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
-
-            @Override
-            public void onCheckedChanged(RadioGroup group, int checkedId) {
-                Log.d(TAG, "Select heading");
-
-                if (checkedId == R.id.headingNext) {
-                    mHeadingMode = WaypointMissionHeadingMode.AUTO;
-                } else if (checkedId == R.id.headingInitDirec) {
-                    mHeadingMode = WaypointMissionHeadingMode.USING_INITIAL_DIRECTION;
-                } else if (checkedId == R.id.headingRC) {
-                    mHeadingMode = WaypointMissionHeadingMode.CONTROL_BY_REMOTE_CONTROLLER;
-                } else if (checkedId == R.id.headingWP) {
-                    mHeadingMode = WaypointMissionHeadingMode.USING_WAYPOINT_HEADING;
-                }
-            }
-        });
-
-        new AlertDialog.Builder(this)
-                .setTitle("")
-                .setView(wayPointSettings)
-                .setPositiveButton("Finish", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-
-                        String altitudeString = wpAltitude_TV.getText().toString();
-                        altitude = Integer.parseInt(nulltoIntegerDefalt(altitudeString));
-                        Log.e(TAG, "altitude " + altitude);
-                        Log.e(TAG, "speed " + mSpeed);
-                        Log.e(TAG, "mFinishedAction " + mFinishedAction);
-                        Log.e(TAG, "mHeadingMode " + mHeadingMode);
-                        configWayPointMission();
-                    }
-
-                })
-                .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        dialog.cancel();
-                    }
-
-                })
-                .create()
-                .show();
-    }
-
-    String nulltoIntegerDefalt(String value) {
-        if (!isIntValue(value)) value = "0";
-        return value;
-    }
-
-    boolean isIntValue(String val) {
-        try {
-            val = val.replace(" ", "");
-            Integer.parseInt(val);
-        } catch (Exception e) {
-            return false;
-        }
-        return true;
-    }
-
-    private void configWayPointMission() {
-
-        if (waypointMissionBuilder == null) {
-
-            waypointMissionBuilder = new WaypointMission.Builder().finishedAction(mFinishedAction)
-                    .headingMode(mHeadingMode)
-                    .autoFlightSpeed(mSpeed)
-                    .maxFlightSpeed(mSpeed)
-                    .flightPathMode(WaypointMissionFlightPathMode.NORMAL);
-
-        } else {
-            waypointMissionBuilder.finishedAction(mFinishedAction)
-                    .headingMode(mHeadingMode)
-                    .autoFlightSpeed(mSpeed)
-                    .maxFlightSpeed(mSpeed)
-                    .flightPathMode(WaypointMissionFlightPathMode.NORMAL);
-
-        }
-
-        if (waypointMissionBuilder.getWaypointList().size() > 0) {
-
-            for (int i = 0; i < waypointMissionBuilder.getWaypointList().size(); i++) {
-                waypointMissionBuilder.getWaypointList().get(i).altitude = altitude;
-            }
-
-            setResultToToast("Set Waypoint attitude successfully");
-        }
-
-        DJIError error = getWaypointMissionOperator().loadMission(waypointMissionBuilder.build());
-        if (error == null) {
-            setResultToToast("loadWaypoint succeeded");
-        } else {
-            setResultToToast("loadWaypoint failed " + error.getDescription());
-        }
-    }
-
-    private void uploadWayPointMission() {
-
-        getWaypointMissionOperator().uploadMission(new CommonCallbacks.CompletionCallback() {
-            @Override
-            public void onResult(DJIError error) {
-                if (error == null) {
-                    setResultToToast("Mission upload successfully!");
-                } else {
-                    setResultToToast("Mission upload failed, error: " + error.getDescription() + " retrying...");
-                    getWaypointMissionOperator().retryUploadMission(null);
-                }
-            }
-        });
-
-    }
-
-    private void startWaypointMission() {
-
-        getWaypointMissionOperator().startMission(new CommonCallbacks.CompletionCallback() {
-            @Override
-            public void onResult(DJIError error) {
-                setResultToToast("Mission Start: " + (error == null ? "Successfully" : error.getDescription()));
-            }
-        });
-    }
-
-    private void stopWaypointMission() {
-
-        getWaypointMissionOperator().stopMission(new CommonCallbacks.CompletionCallback() {
-            @Override
-            public void onResult(DJIError error) {
-                setResultToToast("Mission Stop: " + (error == null ? "Successfully" : error.getDescription()));
-            }
-        });
-
-    }
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-
-        if (gMap == null) {
-            gMap = googleMap;
-            setUpMap();
-        }
-        LatLng shenzhen = new LatLng(22.5362, 113.9454);
-        gMap.addMarker(new MarkerOptions().position(shenzhen).title("Marker in Shenzhen"));
-        gMap.moveCamera(CameraUpdateFactory.newLatLng(shenzhen));
-    }
-
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        setTitle(R.string.settings);
-    }
-}
-
-
-
-/*
-    //@param savedInstanceState Any saved state we are carrying over into the new activity instance
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        Log.d(TAG, "onCreate");
-
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_help);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setDisplayShowHomeEnabled(true);
-
-        // Get the SupportMapFragment and request notification
-        // when the map is ready to be used.
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-
-//        // Display the fragment as the main content.
-//        getFragmentManager().beginTransaction()
-//                .replace(android.R.id.content, new SettingsFragment())
-//                .commit();
-    }
-
-    // Include the OnCreate() method here too, as described above.
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        // Add a marker in Sydney, Australia,
-        // and move the map's camera to the same location.
-        LatLng sydney = new LatLng(-33.852, 151.211);
-        googleMap.addMarker(new MarkerOptions().position(sydney)
-                .title("Marker in Sydney"));
-        googleMap.moveCamera(CameraUpdateFactory.newLatLng(sydney));
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        onBackPressed();
-        return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        super.onBackPressed();
-        setTitle(R.string.settings);
-    }
-
-    @Override
-    public boolean onPreferenceStartScreen(PreferenceFragmentCompat caller, PreferenceScreen preferenceScreen) {
-        Log.d(TAG, "onPreferenceStartScreen");
-//        Log.d(TAG, "callback called to attach the preference sub screen");
-        FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
-        DroneSettingsFragment fragment = new DroneSettingsFragment();
-        Bundle args = new Bundle();
-        args.putString(PreferenceFragmentCompat.ARG_PREFERENCE_ROOT, preferenceScreen.getKey());
-        fragment.setArguments(args);
-        ft.replace(R.id.pref_container, fragment, preferenceScreen.getKey());
-        ft.addToBackStack(preferenceScreen.getKey());
-        ft.commit();
-        setTitle(preferenceScreen.getTitle());
-        return true;
-    }
-}
-*/
