+++ app/src/main/res/layout-land/activity_help.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="#FFFFFF"
-    tools:context="sq.rogue.rosettadrone.settings.HelpActivity">
+    tools:context="sq.rogue.rosettadrone.settings.MapActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
