--- app/src/main/res/xml/ai_selector.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-
-    android:key="ai_prefs"
-    android:title="AI settings"
-    android:layout_width="500dp"
-    android:layout_height="150dp"
-    android:layout_gravity="center"
-    android:background="@color/background_blue"
-    tools:context="sq.rogue.rosettadrone.MainActivity"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/txt_dia"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_margin="20dp"
-        android:text="Select youre AI action"
-        android:textColor="@android:color/white"
-        android:textSize="25dp"
-        android:textStyle="bold" />
-
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:background="@color/background_blue"
-        android:layout_margin="20dp"
-        android:orientation="horizontal">
-
-        <Button
-            android:id="@+id/btn_cable"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:background="@color/background_blue"
-            android:clickable="true"
-            android:text="Cable Tracking"
-            android:textColor="#5DBCD2"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/btn_no"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:layout_marginLeft="5dp"
-            android:clickable="true"
-            android:background="@color/background_blue"
-            android:text="Mast Inspection"
-            android:textColor="#5DBCD2"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/btn_cancel"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:layout_marginLeft="5dp"
-            android:background="@color/background_blue"
-            android:clickable="true"
-            android:text="Cancel"
-            android:textColor="#5DBCD2"
-            android:textStyle="bold" />
-    </LinearLayout>
-
-</LinearLayout>
